requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cors==1.0.1
cycler==0.12.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
future==1.0.0
gast==0.6.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.3.2
namex==0.0.8
narwhals==1.30.0
numpy==2.0.2
openai==0.28.0
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pickle-mixin==1.0.2
pillow==11.1.0
plotly==6.0.0
plotly-express==0.4.1
propcache==0.3.1
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
statsmodels==0.14.4
streamlit==1.43.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.19.0
zope.event==5.0
zope.interface==7.2
